update on borrow global
