+++ vm-circuit/src/chips/execution_chip/instructions.rs
@@ -14,6 +14,7 @@ pub mod add;
 pub mod and;
 pub mod bit_and;
 pub mod bit_or;
+pub mod borrow_global;
 pub mod br_false;
 pub mod br_true;
 pub mod branch;
@@ -30,7 +31,6 @@ pub mod freeze_ref;
 pub mod ge;
 pub mod gt;
 pub mod imm_borrow_field;
-pub mod imm_borrow_global;
 pub mod imm_borrow_loc;
 pub mod ld_false;
 pub mod ld_true;
@@ -44,7 +44,6 @@ pub mod move_loc;
 pub mod move_to;
 pub mod mul;
 pub mod mut_borrow_field;
-pub mod mut_borrow_global;
 pub mod mut_borrow_loc;
 pub mod neq;
 pub mod nop;
