+++ vm-circuit/src/chips/execution_chip/instructions/borrow_global.rs
@@ -1,10 +1,10 @@
 // Copyright (c) zkMove Authors
 
-use crate::chips::execution_chip::instructions::common::LookupBytecode;
+use crate::chips::execution_chip::instructions::common::{LookupBytecode, Word};
 use crate::chips::execution_chip::instructions::Instructions;
 use crate::chips::execution_chip::lookup_tables::{rw_table::RWLookup, LookupsWithCondition};
 use crate::chips::execution_chip::opcode::Opcode;
-use crate::chips::execution_chip::step_chip::StepChipCells;
+use crate::chips::execution_chip::step_chip::{StepChipCells, WORD_CAPACITY};
 use crate::chips::utilities::Expr;
 use crate::witness::execution_steps::ExecutionStep;
 use crate::witness::rw_operations::{RWOperations, RW};
@@ -13,26 +13,32 @@ use halo2_proofs::circuit::Region;
 use halo2_proofs::plonk::{Error, Expression};
 use std::marker::PhantomData;
 
-pub struct MutBorrowGlobal<F: FieldExt> {
+pub struct BorrowGlobal<const MUTABLE: bool, F: FieldExt> {
     _marker: PhantomData<F>,
 }
 
-impl<F: FieldExt> Instructions<F> for MutBorrowGlobal<F> {
+impl<const MUTABLE: bool, F: FieldExt> Instructions<F> for BorrowGlobal<MUTABLE, F> {
     fn configure(
         cells: &StepChipCells<F>,
         constraints: &mut Vec<(&str, Expression<F>)>,
         lookups: &mut LookupsWithCondition<F>,
     ) {
-        let cond = cells.conditions[Opcode::MutBorrowGlobal.index()]
-            .expression
-            .clone();
+        let opcode = if MUTABLE {
+            Opcode::ImmBorrowGlobal
+        } else {
+            Opcode::MutBorrowGlobal
+        };
+        let cond = cells.conditions[opcode.index()].expression.clone();
 
         let pc_expr = cells.pc.expression.clone() - cells.next_pc.expression.clone() + 1.expr();
         let stack_size_expr =
             cells.stack_size.expression.clone() - cells.next_stack_size.expression.clone();
         let frame_index_expr =
             cells.frame_index.expression.clone() - cells.next_frame_index.expression.clone();
-        let gc_expr = cells.gc.expression.clone() - cells.next_gc.expression.clone() + 3.expr();
+        let word_elem_num_expr = cells.auxiliary_3.expression.clone();
+        let gc_expr = cells.gc.expression.clone() - cells.next_gc.expression.clone()
+            + 2.expr()
+            + word_elem_num_expr.clone();
         let module_index =
             cells.module_index.expression.clone() - cells.next_module_index.expression.clone();
         let func_index =
@@ -46,32 +52,36 @@ impl<F: FieldExt> Instructions<F> for MutBorrowGlobal<F> {
             ("function index", cond.clone() * func_index),
         ]);
 
+        let account_address_expr = cells.value_a.expression.clone(); // address
+        let sd_index_expr = cells.auxiliary_1.expression.clone(); //sd_index
         lookups.rw_lookups.push((
             RWLookup::stack_pop(
                 cells.gc.expression.clone(),
                 cells.stack_size.expression.clone(),
                 0.expr(),
                 0.expr(),
-                cells.value_a.expression.clone(), //address
+                account_address_expr.clone(),
             ),
             cond.clone(),
         ));
 
-        lookups.rw_lookups.push((
-            RWLookup::global_read(
-                cells.gc.expression.clone() + 1.expr(),
-                cells.value_a.expression.clone(), //address
-                cells.value_b.expression.clone(),
-                cells.auxiliary_1.expression.clone(), //sd_index
-                0.expr(),
-                0.expr(),
-            ),
-            cond.clone(),
-        ));
+        for i in 0..WORD_CAPACITY {
+            lookups.rw_lookups.push((
+                RWLookup::global_read(
+                    cells.gc.expression.clone() + (i as u64 + 1).expr(),
+                    account_address_expr.clone(),
+                    cells.word_a[i].expression.clone(),
+                    sd_index_expr.clone(),
+                    cells.word_a_addr_ext_0[i].expression.clone(),
+                    cells.word_a_addr_ext_1[i].expression.clone(),
+                ),
+                cond.clone() * (1.expr() - cells.word_a_mask[i].expression.clone()),
+            ));
+        }
 
         lookups.rw_lookups.push((
             RWLookup::stack_push(
-                cells.gc.expression.clone() + 2.expr(),
+                cells.gc.expression.clone() + 1.expr() + word_elem_num_expr,
                 cells.stack_size.expression.clone() - 1.expr(),
                 0.expr(),
                 0.expr(),
@@ -80,10 +90,13 @@ impl<F: FieldExt> Instructions<F> for MutBorrowGlobal<F> {
             cond.clone(),
         ));
 
+        // TODO: constraint reference(cells.value_c) with account_address/sd_index
+        // see issue https://github.com/young-rocks/zkmove-vm/issues/78
+
         LookupBytecode::lookup_bytecode(
             cells,
-            Opcode::MutBorrowGlobal,
-            cells.auxiliary_1.expression.clone(),
+            opcode,
+            sd_index_expr,
             &mut lookups.bytecode_lookups,
             cond,
         );
@@ -100,17 +113,28 @@ impl<F: FieldExt> Instructions<F> for MutBorrowGlobal<F> {
         debug_assert!(op.rw() == RW::READ);
         cells.value_a.assign(region, offset, op.value().value())?;
 
-        let op = rw_operations.0.get(step.gc + 1).ok_or(Error::Synthesis)?;
-        debug_assert!(op.rw() == RW::READ);
-        cells.value_b.assign(region, offset, op.value().value())?;
         cells
             .auxiliary_1
             .assign(region, offset, Some(F::from_u128(op.sd_index() as u128)))?;
 
-        let op = rw_operations.0.get(step.gc + 2).ok_or(Error::Synthesis)?;
+        let word_elem_num = Word::get_word_element_num(region, offset, step, cells)?;
+        Word::assign_word_a(
+            region,
+            offset,
+            step,
+            rw_operations,
+            cells,
+            step.gc + 1,
+            word_elem_num,
+        )?;
+
+        // reference written to stack
+        let op = rw_operations
+            .0
+            .get(step.gc + 1 + word_elem_num)
+            .ok_or(Error::Synthesis)?;
         debug_assert!(op.rw() == RW::WRITE);
         cells.value_c.assign(region, offset, op.value().value())?;
-
         Ok(())
     }
 }
